Start a bootstrapper that generates a Horde development environment based on git + PHP + the basic PEAR package.
It should be sufficient to run

 php components/bin/horde-bootstrap

in the horde checkout to install the development environment in

 lib/pear

The command is not yet able to update the environment which still needs to be implemented.

Right now the installation is still network dependent but it would be a future target to allow it to run from a collection of defined PEAR package archives.