Add missing files (Bug #10296).
components/test/Components/fixture/.gitignore is the culprit that hid
these files on my local system. In principle it would be better to
rename the .gitignore to something that is not parsed by git. This
would require making this an optional argument just for the sake of
testing though. And I consider adding these behind the back of the
.gitignore the easier solution for now.