Introduce a doc/changelog.yml file.
This will be the authoritative source for all changes across all branches. CHANGES and package.xml will be built from this source.
The idea is to keep this changelog.yml file the same in all branches, so that it's much easier to merge/cherry-pick commits including the changelog from one branch to the other without conflicts.
Beside that, it's easier to amend a changelog and update all the other changelogs with a single horde-components call.
Finally, this will allow us to automatically build chnangelogs for bundles in the future.